requirements.txt
Django==1.4.3
dj-database-url==0.2.1
django-crispy-forms==1.2.3
django-form-utils==0.2.0
django-registration==0.8
django-selectable==0.6.2
psycopg2==2.4.5
python-dateutil==2.1
pytz==2012j
six==1.2.0
wsgiref==0.1.2
